+++ backend/pcloud-api/src/test/java/com/api/exhibition/application/ExhibitionServiceTest.java
@@ -4,11 +4,11 @@ import com.api.exhibition.application.dto.ExhibitionCreateRequest;
 import com.api.exhibition.application.dto.ExhibitionUpdateRequest;
 import com.common.exception.AuthException;
 import com.common.exception.AuthExceptionType;
-import com.domain.exhibition.domain.Exhibition;
-import com.domain.exhibition.domain.ExhibitionRepository;
-import com.domain.exhibition.exception.ExhibitionException;
-import com.domain.exhibition.exception.ExhibitionExceptionType;
-import exhibition.infrasturcture.ExhibitionFakeRepository;
+import com.domain.show.exhibition.domain.Exhibition;
+import com.domain.show.exhibition.domain.ExhibitionRepository;
+import com.domain.show.exhibition.exception.ExhibitionException;
+import com.domain.show.exhibition.exception.ExhibitionExceptionType;
+import show.exhibition.infrasturcture.ExhibitionFakeRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -18,8 +18,8 @@ import org.junit.jupiter.api.Test;
 
 import static com.api.exhibition.fixture.ExhibitionRequestFixtures.개인전시회_생성_요청_생성;
 import static com.api.exhibition.fixture.ExhibitionRequestFixtures.개인전시회_업데이트_요청_생성;
-import static exhibition.domain.ExhibitionFixture.개인전시회_생성_사진_개인전;
-import static exhibition.domain.LikedExhibitionFixture.개인_전시회_좋아요_생성_개인전시회아이디_회원아이디;
+import static show.exhibition.domain.ExhibitionFixture.개인전시회_생성_사진_개인전;
+import static show.exhibition.domain.LikedExhibitionFixture.개인_전시회_좋아요_생성_개인전시회아이디_회원아이디;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
