+++ backend/pcloud-api/src/test/java/com/api/exhibition/fixture/ExhibitionRequestFixtures.java
@@ -2,7 +2,7 @@ package com.api.exhibition.fixture;
 
 import com.api.exhibition.application.dto.ExhibitionCreateRequest;
 import com.api.exhibition.application.dto.ExhibitionUpdateRequest;
-import com.domain.common.PublicTag;
+import com.domain.show.common.PublicTag;
 import java.time.LocalDateTime;
 import java.util.List;
