+++ backend/pcloud-api/src/test/java/com/api/exhibition/presentation/ExhibitionControllerWebMvcTest.java
@@ -3,8 +3,8 @@ package com.api.exhibition.presentation;
 import com.api.exhibition.application.dto.ExhibitionCreateRequest;
 import com.api.exhibition.application.dto.ExhibitionUpdateRequest;
 import com.api.helper.MockBeanInjection;
-import com.domain.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
-import com.domain.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
+import com.domain.show.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
+import com.domain.show.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Optional;
@@ -18,8 +18,8 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static com.api.exhibition.fixture.ExhibitionRequestFixtures.개인전시회_생성_요청_생성;
 import static com.api.exhibition.fixture.ExhibitionRequestFixtures.개인전시회_업데이트_요청_생성;
-import static exhibition.infrasturcture.ExhibitionSimpleResponseFixture.개인전시회_간단_조회_응답_생성;
-import static exhibition.infrasturcture.ExhibitionSpecificResponseFixture.개인전시회_상세_조회_응답_생성;
+import static show.exhibition.domain.ExhibitionSimpleResponseFixture.개인전시회_간단_조회_응답_생성;
+import static show.exhibition.domain.ExhibitionSpecificResponseFixture.개인전시회_상세_조회_응답_생성;
 import static com.api.helper.RestDocsHelper.customDocument;
 import static member.fixture.MemberFixture.어드민_멤버_생성_id_없음_kakao_oauth_가입;
 import static org.mockito.ArgumentMatchers.anyInt;
