+++ backend/pcloud-api/src/test/java/com/api/popups/application/PopupsServiceTest.java
@@ -3,26 +3,26 @@ package com.api.popups.application;
 import com.api.popups.application.request.PopupsCreateRequest;
 import com.common.exception.AuthException;
 import com.common.exception.AuthExceptionType;
-import com.domain.popups.domain.Popups;
-import com.domain.popups.domain.PopupsRepository;
-import com.domain.popups.exception.PopupsException;
+import com.domain.show.popups.domain.Popups;
+import com.domain.show.popups.domain.PopupsRepository;
+import com.domain.show.popups.exception.PopupsException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import popups.FakePopupsRepository;
+import show.popups.infrastructure.FakePopupsRepository;
 
 import java.util.Optional;
 
 import static com.api.popups.fixture.request.PopupsRequestFixtures.팝업스토어_생성_요청;
-import static com.domain.popups.exception.PopupsExceptionType.POPUPS_NOT_FOUND_EXCEPTION;
+import static com.domain.show.popups.exception.PopupsExceptionType.POPUPS_NOT_FOUND_EXCEPTION;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static popups.fixture.PopupsFixture.일반_팝업_스토어_생성_뷰티;
-import static popups.fixture.PopupsFixture.일반_팝업_스토어_생성_뷰티_유효하지_않은_주인;
-import static popups.fixture.PopupsFixture.일반_팝업_스토어_생성_펫샵;
+import static show.popups.domain.PopupsFixture.일반_팝업_스토어_생성_뷰티;
+import static show.popups.domain.PopupsFixture.일반_팝업_스토어_생성_뷰티_유효하지_않은_주인;
+import static show.popups.domain.PopupsFixture.일반_팝업_스토어_생성_펫샵;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
