+++ backend/pcloud-api/src/test/java/com/api/popups/fixture/request/PopupsRequestFixtures.java
@@ -2,7 +2,7 @@ package com.api.popups.fixture.request;
 
 import com.api.popups.application.request.PopupsCreateRequest;
 import com.api.popups.application.request.PopupsUpdateRequest;
-import com.domain.common.PublicTag;
+import com.domain.show.common.PublicTag;
 
 import java.time.LocalDateTime;
 import java.util.List;
