--- backend/pcloud-domain/src/testFixtures/java/popups/fixture/LikedPopupsFixture.java
@@ -1,10 +1,0 @@
-package popups.fixture;
-
-import com.domain.popups.domain.LikedPopups;
-
-public class LikedPopupsFixture {
-
-    public static LikedPopups 팝업_좋아요_생성_팝업_아이디_1_멤버_아이디_1() {
-        return LikedPopups.of(1L, 1L);
-    }
-}
