+++ backend/pcloud-domain/src/testFixtures/java/show/popups/domain/LikedPopupsFixture.java
@@ -1,0 +1,10 @@
+package show.popups.domain;
+
+import com.domain.show.popups.domain.LikedPopups;
+
+public class LikedPopupsFixture {
+
+    public static LikedPopups 팝업_좋아요_생성_팝업_아이디_1_멤버_아이디_1() {
+        return LikedPopups.of(1L, 1L);
+    }
+}
